Split dashboard map preview hooks
